+++ CHANGELOG.md
@@ -12,6 +12,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `failure` support must be activated via feature `with-failure`
 - If `failure` feature is activated, both `HttpApiProblem` and `ApiError` implement `failure::Fail`, otherwise `std::error::Error`
 - Feature `with_warp` became `with-warp`
-- Feature `with_actix` became `with-warp`
-- Feature `with_warp` became `with-warp`
-- Feature `with_warp` became `with-warp`
+- Feature `with_actix_web` became `with-actix-web`
+- Feature `with_hyper` became `with-hyper`